Moved enriching testing to test_enriching.py module. Found bug in enrich_with_weather_data and fixed it
