+++ tests/processing/test_post_process.py
@@ -4,7 +4,7 @@ from src.processing.post_process import generate_centres_df
 from src.processing.pre_process import df_cleaner, df_generator, df_group_and_filter
 
 
-def test_df_generator_read_csv(get_path):
+def test_generate_centres_df(get_path):
     data_frames = df_generator(get_path + "/tests/test_data/hotels_test_data.zip")
     df_full = df_group_and_filter([df_cleaner(df) for df in data_frames])
     centre_df = generate_centres_df(df_full)
