+++ tests/processing/test_pre_process.py
@@ -1,15 +1,8 @@
 # -*- coding: utf-8 -*-
 
-from unittest.mock import patch
-
 import pytest
 
-from src.processing.enriching import enrich_with_geo_data
 from src.processing.pre_process import df_cleaner, df_generator, df_group_and_filter
-
-
-def fake_get_address_worker(coordinate):
-    return "Totally real address of real Country"
 
 
 def test_df_generator_read_csv(get_path):
@@ -38,13 +31,3 @@ def test_df_group_and_filter(get_path):
     df_full = df_group_and_filter([df_cleaner(df) for df in data_frames])
 
     assert df_full["City"].values[0] == "Oak Brook"
-
-
-def test_enrich_with_geo_data(get_path):
-    data_frames = df_generator(get_path + "/tests/test_data/hotels_test_data.zip")
-    df_full = df_group_and_filter([df_cleaner(df) for df in data_frames])
-
-    with patch("src.api_utils.geodata_api.get_address_worker", fake_get_address_worker):
-        enrich_with_geo_data(df_full)
-
-    assert df_full["Address"].values[0] == "Totally real address of real Country"
