+++ tests/test_data/weather_api_historical_worker_sample_resp.json
@@ -1,0 +1,1 @@
+{"lat": 52.3376, "lon": 4.8178, "timezone": "Europe/Amsterdam", "timezone_offset": 3600, "current": {"dt": 1643032817, "sunrise": 1643009617, "sunset": 1643040697, "temp": 5.17, "feels_like": 5.17, "pressure": 1034, "humidity": 75, "dew_point": 1.11, "uvi": 0.23, "clouds": 75, "visibility": 10000, "wind_speed": 0.45, "wind_deg": 329, "wind_gust": 1.79, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04d"}]}, "hourly": [{"dt": 1642982400, "temp": 5.36, "feels_like": 3.73, "pressure": 1035, "humidity": 81, "dew_point": 2.37, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 2.06, "wind_deg": 220, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1642986000, "temp": 5.26, "feels_like": 4.19, "pressure": 1034, "humidity": 81, "dew_point": 2.27, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 1.54, "wind_deg": 140, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1642989600, "temp": 5.2, "feels_like": 4.12, "pressure": 1034, "humidity": 84, "dew_point": 2.72, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 1.54, "wind_deg": 140, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1642993200, "temp": 5.15, "feels_like": 3.48, "pressure": 1034, "humidity": 83, "dew_point": 2.5, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 2.06, "wind_deg": 140, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1642996800, "temp": 5.09, "feels_like": 5.09, "pressure": 1034, "humidity": 81, "dew_point": 2.1, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 0.51, "wind_deg": 0, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1643000400, "temp": 5.68, "feels_like": 4.67, "pressure": 1034, "humidity": 79, "dew_point": 2.33, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 1.54, "wind_deg": 140, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1643004000, "temp": 4.98, "feels_like": 3.87, "pressure": 1034, "humidity": 83, "dew_point": 2.34, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 1.54, "wind_deg": 160, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1643007600, "temp": 4.94, "feels_like": 3.24, "pressure": 1034, "humidity": 83, "dew_point": 2.3, "uvi": 0, "clouds": 90, "visibility": 10000, "wind_speed": 2.06, "wind_deg": 140, "weather": [{"id": 804, "main": "Clouds", "description": "overcast clouds", "icon": "04n"}]}, {"dt": 1643011200, "temp": 4.9, "feels_like": 3.19, "pressure": 1034, "humidity": 81, "dew_point": 1.92, "uvi": 0, "clouds": 90, "visibility": 9000, "wind_speed": 2.06, "wind_deg": 140, "weather": [{"id": 804, "main": "Clouds", "description": "overcast clouds", "icon": "04d"}]}, {"dt": 1643014800, "temp": 4.99, "feels_like": 4.99, "pressure": 1034, "humidity": 80, "dew_point": 1.83, "uvi": 0.24, "clouds": 90, "visibility": 8000, "wind_speed": 0.45, "wind_deg": 329, "wind_gust": 1.79, "weather": [{"id": 804, "main": "Clouds", "description": "overcast clouds", "icon": "04d"}]}, {"dt": 1643018400, "temp": 5.01, "feels_like": 3.32, "pressure": 1032, "humidity": 77, "dew_point": 1.32, "uvi": 0.48, "clouds": 90, "visibility": 9000, "wind_speed": 2.06, "wind_deg": 170, "weather": [{"id": 804, "main": "Clouds", "description": "overcast clouds", "icon": "04d"}]}, {"dt": 1643022000, "temp": 5.04, "feels_like": 5.04, "pressure": 1033, "humidity": 75, "dew_point": 0.98, "uvi": 0.68, "clouds": 90, "visibility": 10000, "wind_speed": 0.45, "wind_deg": 329, "wind_gust": 1.79, "weather": [{"id": 804, "main": "Clouds", "description": "overcast clouds", "icon": "04d"}]}, {"dt": 1643025600, "temp": 5.08, "feels_like": 5.08, "pressure": 1034, "humidity": 75, "dew_point": 1.02, "uvi": 0.72, "clouds": 90, "visibility": 10000, "wind_speed": 1.03, "wind_deg": 0, "weather": [{"id": 804, "main": "Clouds", "description": "overcast clouds", "icon": "04d"}]}, {"dt": 1643029200, "temp": 5.13, "feels_like": 5.13, "pressure": 1034, "humidity": 75, "dew_point": 1.07, "uvi": 0.38, "clouds": 90, "visibility": 10000, "wind_speed": 1.03, "wind_deg": 0, "weather": [{"id": 804, "main": "Clouds", "description": "overcast clouds", "icon": "04d"}]}, {"dt": 1643032800, "temp": 5.17, "feels_like": 5.17, "pressure": 1034, "humidity": 75, "dew_point": 1.11, "uvi": 0.23, "clouds": 75, "visibility": 10000, "wind_speed": 0.45, "wind_deg": 329, "wind_gust": 1.79, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04d"}]}, {"dt": 1643036400, "temp": 5.07, "feels_like": 5.07, "pressure": 1033, "humidity": 76, "dew_point": 1.2, "uvi": 0.09, "clouds": 75, "visibility": 10000, "wind_speed": 0.45, "wind_deg": 329, "wind_gust": 1.34, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04d"}]}, {"dt": 1643040000, "temp": 4.77, "feels_like": 4.77, "pressure": 1033, "humidity": 76, "dew_point": 0.9, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 1.03, "wind_deg": 0, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04d"}]}, {"dt": 1643043600, "temp": 4.3, "feels_like": 3.1, "pressure": 1034, "humidity": 78, "dew_point": 0.81, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 1.54, "wind_deg": 130, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1643047200, "temp": 4.23, "feels_like": 3.02, "pressure": 1034, "humidity": 79, "dew_point": 0.92, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 1.54, "wind_deg": 80, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1643050800, "temp": 4.17, "feels_like": 4.17, "pressure": 1034, "humidity": 80, "dew_point": 1.03, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 0.45, "wind_deg": 329, "wind_gust": 1.34, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1643054400, "temp": 4.05, "feels_like": 2.2, "pressure": 1035, "humidity": 80, "dew_point": 0.92, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 2.06, "wind_deg": 110, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1643058000, "temp": 3.95, "feels_like": 1.6, "pressure": 1033, "humidity": 78, "dew_point": 0.47, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 2.57, "wind_deg": 100, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1643061600, "temp": 3.93, "feels_like": 2.67, "pressure": 1033, "humidity": 78, "dew_point": 0.45, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 1.54, "wind_deg": 80, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}, {"dt": 1643065200, "temp": 3.86, "feels_like": 3.86, "pressure": 1032, "humidity": 78, "dew_point": 0.38, "uvi": 0, "clouds": 75, "visibility": 10000, "wind_speed": 1.03, "wind_deg": 0, "weather": [{"id": 803, "main": "Clouds", "description": "broken clouds", "icon": "04n"}]}]}
